move ‘string(for floorChangeType:)’, ‘string(forTurn direction:)’, and ‘string(forFloorChangeDirection:)’ to an extension on InstructionViewModel
